+++ app/db/migrate/20250516133131_deposite_money.rb
@@ -2,7 +2,7 @@ class DepositeMoney < ActiveRecord::Migration[8.0]
   def change
     create_table :deposite_money do |t|
       t.integer :operation_number
-      t.float :amount
+      t.integer :amount
       t.date :date_money_deposited
     end
   end
